chore: refactor batch logic for reuse
