+++ src/dynamodb/batch/batch-delete.test.js
@@ -1,0 +1,5 @@
+describe('batch delete', () => {
+  it('should work', () => {
+    expect(true).toBeTruthy();
+  });
+});
